My first pass through the book, focusing on the Ship of Theseus text only.
Enjoyed the story, did my best to ignore the marginalia and footnotes. It was interesting to read just the SOT text and then get snatches of intrigue and excitement from experiencing Jen and Eric's story in a nonlinear manner. It was also hard to stop myself from starting to delve into the codes and artifacts from the book - looking forward to pass 2.