Second pass through the book
This time focusing on Jennifer and Eric's footnotes and the artifacts that they reference in their margin notes. It was still hard to avoid getting caught up in the coded messages from FXC in the footnotes (and chapter text??) - I was actually kind of annoyed that the margin notes spoil the codes in several chapters.

I think my third pass will probably be a codebreaking and reference pass to try to nail down more of the people, dates, and codes. Also, the raised letter code in one of the later chapters is **obvious**. Difficult to believe that Jennifer/Eric wouldn't have noticed it.